requirements.txt
Django==3.2.5
psycopg2==2.9.1
requests==2.26.0
python-dateutil
django-tempus-dominus
django-mptt
django-recaptcha
tldextract
Pillow
bs4
gunicorn
pytz